Add custom signing function to AppleWalletGenerator class
This change was made so those who use this package can pass in a
function from their project that handles signing the manifest.json
file.

Apple requires signing for all wallet cards. To do this with the
current package logic would require adding certificates/private keys
to the app bundle which is not the most secure approach.

Instead, I have updated our private API with the signing logic and
pass in the function from my app project that calls our API to do
the signing.